+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Added underline to instance name on welcome screen.
+
 ## [3.18.0] - 2025-02-26
 
 ### Added