+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added underline to instance name on welcome screen.
 - Reduced transparency of image description input.
+- Changed "Remove attachment" button style.
 
 ## [3.18.0] - 2025-02-26