+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added underline to instance name on welcome screen.
 - Reduced transparency of image description input.
 - Changed "Remove attachment" button style.
+- Replaced "Click here to edit description" button with an icon.
 
 ## [3.18.0] - 2025-02-26