+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Replaced "Click here to edit description" button with an icon.
 - Remove meta tags inserted by server.
 
+### Fixed
+
+- Add empty `assets/custom.css` to prevent error 404.
+
 ## [3.18.0] - 2025-02-26
 
 ### Added