+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Replaced "Click here to edit description" button with an icon.
 - Remove meta tags inserted by server.
 - Inject custom CSS link tag at build time.
+- Updated translations ([#68](https://codeberg.org/silverpill/mitra-web/pulls/68)).
 
 ### Fixed