+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support generating identity proofs with `signify` tool.
+
 ## [3.19.0] - 2025-03-09
 
 ### Changed