+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [3.20.0] - 2025-03-15
+
 ### Added
 
 - Support generating identity proofs with `signify` tool.